+++ replacer-backend/replacer-web/src/main/java/es/bvalero/replacer/page/save/ReviewSaveController.java
@@ -81,7 +81,7 @@ public class ReviewSaveController {
                 .queryTimestamp(queryTimestamp)
                 .build();
 
-            reviewSaveService.saveReviewContent(pageSave, user.getAccessToken());
+            reviewSaveService.saveReviewContent(pageSave, user);
             reviewSaveService.markAsReviewed(reviewedReplacements, true);
         }