+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.25.0 (??)
 - Feature: find replacements for Spanish ordinals with wrong format
+- Feature: additional check in server before saving a page that the user has not reached the maximum number of editions per minute
+- Fix: display the correct error message in the alert
 
 ## 2.24.2 (24 Sep 2023)
 - Fix regression: replacements not marked as reviewed when they are contained in a section